+++ packages/stream_video/lib/src/call/call.dart
@@ -1141,6 +1141,7 @@ class Call {
     StreamBroadcastingSettings? broadcasting,
     StreamSessionSettings? session,
     StreamFrameRecordingSettings? frameRecording,
+    StreamIngressSettings? ingress,
   }) {
     return _coordinatorClient.updateCall(
       callCid: callCid,
@@ -1158,6 +1159,7 @@ class Call {
       broadcasting: broadcasting,
       session: session,
       frameRecording: frameRecording,
+      ingress: ingress,
     );
   }
 
@@ -2161,6 +2163,7 @@ class Call {
     StreamBroadcastingSettings? broadcasting,
     StreamGeofencingSettings? geofencing,
     StreamSessionSettings? session,
+    StreamIngressSettings? ingress,
     StreamFrameRecordingSettings? frameRecording,
     Map<String, Object> custom = const {},
   }) async {
@@ -2176,6 +2179,7 @@ class Call {
       broadcasting: broadcasting?.toOpenDto(),
       geofencing: geofencing?.toOpenDto(),
       session: session?.toOpenDto(),
+      ingress: ingress?.toOpenDto(),
       frameRecording: frameRecording?.toOpenDto(),
     );
 
@@ -2247,6 +2251,15 @@ class Call {
 
   Future<Result<None>> unblockUser(String userId) {
     return _permissionsManager.unblockUser(userId);
+  }
+
+  /// Kicks a user from the call.
+  /// Set [block] to true to also block the user from rejoining.
+  Future<Result<None>> kickUser(
+    String userId, {
+    bool block = false,
+  }) {
+    return _permissionsManager.kickUser(userId, block: block);
   }
 
   Future<Result<None>> startRecording({
@@ -2628,14 +2641,19 @@ class Call {
         }).asCancelable(),
       );
 
-      // Set multitasking camera access for iOS
-      final multitaskingResult = await setMultitaskingCameraAccessEnabled(
-        enabled && !_streamVideo.options.muteVideoWhenInBackground,
-      );
+      var multitaskingEnabled = state.value.iOSMultitaskingCameraAccessEnabled;
+      if (enabled && !multitaskingEnabled) {
+        // Set multitasking camera access for iOS
+        final multitaskingResult = await setMultitaskingCameraAccessEnabled(
+          enabled && !_streamVideo.options.muteVideoWhenInBackground,
+        );
+
+        multitaskingEnabled = multitaskingResult.getDataOrNull() ?? false;
+      }
 
       _stateManager.participantSetCameraEnabled(
         enabled: enabled,
-        iOSMultitaskingCameraAccessEnabled: multitaskingResult.getDataOrNull(),
+        iOSMultitaskingCameraAccessEnabled: multitaskingEnabled,
       );
 
       _connectOptions = _connectOptions.copyWith(