+++ packages/stream_video/lib/src/coordinator/open_api/event/open_api_event.dart
@@ -30,6 +30,7 @@ class OpenApiEvent with EquatableMixin {
     this.callPermissionsUpdated,
     this.callUserBlocked,
     this.callUserUnblocked,
+    this.callUserKicked,
     this.callRecordingStarted,
     this.callRecordingStopped,
     this.callBroadcastingStarted,
@@ -121,6 +122,9 @@ class OpenApiEvent with EquatableMixin {
       case EventType.callUserUnblocked:
         final event = open.UnblockedUserEvent.fromJson(jsonObj);
         return result.copyWith(callUserUnblocked: event);
+      case EventType.callUserKicked:
+        final event = open.KickedUserEvent.fromJson(jsonObj);
+        return result.copyWith(callUserKicked: event);
       case EventType.callRecordingStarted:
         final event = open.CallRecordingStartedEvent.fromJson(jsonObj);
         return result.copyWith(callRecordingStarted: event);
@@ -235,6 +239,7 @@ class OpenApiEvent with EquatableMixin {
   final open.UpdatedCallPermissionsEvent? callPermissionsUpdated;
   final open.BlockedUserEvent? callUserBlocked;
   final open.UnblockedUserEvent? callUserUnblocked;
+  final open.KickedUserEvent? callUserKicked;
   final open.CallRecordingStartedEvent? callRecordingStarted;
   final open.CallRecordingStoppedEvent? callRecordingStopped;
   final open.CallHLSBroadcastingStartedEvent? callBroadcastingStarted;
@@ -285,6 +290,7 @@ class OpenApiEvent with EquatableMixin {
     open.UpdatedCallPermissionsEvent? callPermissionsUpdated,
     open.BlockedUserEvent? callUserBlocked,
     open.UnblockedUserEvent? callUserUnblocked,
+    open.KickedUserEvent? callUserKicked,
     open.CallRecordingStartedEvent? callRecordingStarted,
     open.CallRecordingStoppedEvent? callRecordingStopped,
     open.CallHLSBroadcastingStartedEvent? callBroadcastingStarted,
@@ -339,6 +345,7 @@ class OpenApiEvent with EquatableMixin {
           callPermissionsUpdated ?? this.callPermissionsUpdated,
       callUserBlocked: callUserBlocked ?? this.callUserBlocked,
       callUserUnblocked: callUserUnblocked ?? this.callUserUnblocked,
+      callUserKicked: callUserKicked ?? this.callUserKicked,
       callRecordingStarted: callRecordingStarted ?? this.callRecordingStarted,
       callRecordingStopped: callRecordingStopped ?? this.callRecordingStopped,
       callBroadcastingStarted:
@@ -408,6 +415,7 @@ class OpenApiEvent with EquatableMixin {
         callPermissionsUpdated,
         callUserBlocked,
         callUserUnblocked,
+        callUserKicked,
         callRecordingStarted,
         callRecordingStopped,
         callBroadcastingStarted,