+++ packages/stream_video/lib/src/coordinator/retry/coordinator_client_retry.dart
@@ -52,6 +52,24 @@ class CoordinatorClientRetry extends CoordinatorClient {
   }
 
   @override
+  Future<Result<None>> kickUser({
+    required StreamCallCid callCid,
+    required String userId,
+    bool block = false,
+  }) {
+    return _retryManager.execute(
+      () => _delegate.kickUser(
+        callCid: callCid,
+        userId: userId,
+        block: block,
+      ),
+      (error, nextAttemptDelay) async {
+        _logRetry('kickUser', error, nextAttemptDelay);
+      },
+    );
+  }
+
+  @override
   Future<Result<None>> createDevice({
     required String id,
     required PushProvider pushProvider,
@@ -679,6 +697,7 @@ class CoordinatorClientRetry extends CoordinatorClient {
     StreamBroadcastingSettings? broadcasting,
     StreamSessionSettings? session,
     StreamFrameRecordingSettings? frameRecording,
+    StreamIngressSettings? ingress,
   }) {
     return _retryManager.execute(
       () => _delegate.updateCall(
@@ -696,6 +715,7 @@ class CoordinatorClientRetry extends CoordinatorClient {
         limits: limits,
         session: session,
         frameRecording: frameRecording,
+        ingress: ingress,
       ),
       (error, nextAttemptDelay) async {
         _logRetry('updateCall', error, nextAttemptDelay);