+++ packages/stream_video/lib/src/sorting/call_participant_state_sorting.dart
@@ -108,7 +108,7 @@ Comparator<CallParticipantState> byRole(List<String> roles) {
     );
 
     if (aMatches > bMatches) return -1;
-    if (bMatches < aMatches) return 1;
+    if (aMatches < bMatches) return 1;
     return 0;
   };
 }